requirements.txt
numpy
pandas
requests
PyYAML
statsmodels
scipy
scikit-learn
ruptures
dtaidistance
tabulate
